+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 5.16.0 (planned to 2x.10.2020)
 * #1051 Selenide plugins system
+* #1261 Add actual own text to error message (when one of checks `ownText`, `exactOwnText` fails)
 
 ## 5.15.1 (released 03.10.2020)
 * Fix creating logs dir in parallel tests